% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access<-}
\alias{access<-}
\title{Write access population or transition arrays generated by JHEEM by dimension names}
\usage{
access(
  arr,
  age = NULL,
  race = NULL,
  subpopulation = NULL,
  sex = NULL,
  risk = NULL,
  non.hiv.subset = NULL,
  continuum = NULL,
  cd4 = NULL,
  hiv.subset = NULL,
  age.from = NULL,
  race.from = NULL,
  subpopulation.from = NULL,
  sex.from = NULL,
  risk.from = NULL,
  non.hiv.subset.from = NULL,
  continuum.from = NULL,
  cd4.from = NULL,
  hiv.subset.from = NULL,
  age.to = NULL,
  race.to = NULL,
  subpopulation.to = NULL,
  sex.to = NULL,
  risk.to = NULL,
  non.hiv.subset.to = NULL,
  continuum.to = NULL,
  cd4.to = NULL,
  hiv.subset.to = NULL,
  year = NULL
) <- value
}
\arguments{
\item{arr}{An array generated by a \code{\link{jheem-class}} object, usually by a call to get.XXX.skeleton (eg  \code{\link{get.population.skeleton}} or \code{\link{get.transition.array.skeleton.hiv.negative}})}

\item{age}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{race}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{subpopulation}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{sex}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{risk}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{non.hiv.subset}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{continuum}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{cd4}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{hiv.subset}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{age.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{race.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{subpopulation.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{sex.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{risk.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{non.hiv.subset.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{continuum.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{cd4.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{hiv.subset.from}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{age.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{race.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{subpopulation.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{sex.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{risk.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{non.hiv.subset.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{continuum.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{cd4.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{hiv.subset.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{value}{The value to overwrite the array subset with. If dimnames(value) is not null, access<- will overwrite the values in arr such that the dimnames of the subset of arr being overwritten match with the dimnames of value}
}
\description{
Write access population or transition arrays generated by JHEEM by dimension names
}
\seealso{
Other functions to access JHEEM arrays: 
\code{\link{access}()}
}
\concept{functions to access JHEEM arrays}
