% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\title{Read access population or transition arrays generated by JHEEM by dimension names}
\usage{
access(
  arr,
  age = NULL,
  race = NULL,
  subpopulation = NULL,
  sex = NULL,
  risk = NULL,
  non.hiv.subset = NULL,
  continuum = NULL,
  cd4 = NULL,
  hiv.subset = NULL,
  age.from = NULL,
  race.from = NULL,
  subpopulation.from = NULL,
  sex.from = NULL,
  risk.from = NULL,
  non.hiv.subset.from = NULL,
  continuum.from = NULL,
  cd4.from = NULL,
  hiv.subset.from = NULL,
  age.to = NULL,
  race.to = NULL,
  subpopulation.to = NULL,
  sex.to = NULL,
  risk.to = NULL,
  non.hiv.subset.to = NULL,
  continuum.to = NULL,
  cd4.to = NULL,
  hiv.subset.to = NULL,
  year = NULL,
  collapse.length.one.dimensions = T
)
}
\arguments{
\item{arr}{An array generated by a \code{\link{jheem-class}} object, usually by a call to get.XXX.skeleton (eg  \code{\link{get.population.skeleton}} or \code{\link{get.transition.array.skeleton.hiv.negative}})}

\item{age, race, subpopulation, sex, risk, non.hiv.subset, continuum, cd4, hiv.subset, age.from, race.from, subpopulation.from, sex.from, risk.from, non.hiv.subset.from, continuum.from, cd4.from, hiv.subset.from, age.to, race.to, subpopulation.to, sex.to, risk.to, non.hiv.subset.to, continuum.to, cd4.to, hiv.subset.to}{For each possible array dimension, the values within that dimension (either numeric indices or character names) to subset}

\item{collapse.length.one.dimensions}{If false, even if a dimension in the return value only has length one, it will be preserved in the return value. If true, length one dimensions are removed (as with normal array subsetting)}
}
\description{
Read access population or transition arrays generated by JHEEM by dimension names
}
\seealso{
Other functions to access JHEEM arrays: 
\code{\link{access<-}()}
}
\concept{functions to access JHEEM arrays}
