% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{combine.jheem.results}
\alias{combine.jheem.results}
\title{Combine Two JHEEM Results Objects into One by year}
\usage{
combine.jheem.results(results1, results2)
}
\arguments{
\item{results1, results2}{Two objects of class jheem.results. If years from the two overlap, data from results1 will be used}
}
\description{
Combine Two JHEEM Results Objects into One by year
}
