% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{create.birth.proportions.from.population}
\alias{create.birth.proportions.from.population}
\title{Create a Birth Proportions array based on sex distribution by race from a given population}
\usage{
create.birth.proportions.from.population(
  jheem,
  population,
  risk.born.into.state = 1,
  non.hiv.subset.born.into.state = 1
)
}
\arguments{
\item{jheem}{A JHEEM object}

\item{population}{The population based on which to determine sex distribution by race. Must have at least a dimension for sex}
}
\description{
Creates a birth proportions array with the marginals of sex by race drawn from the population. The returned value is valid as birth proportions for either HIV-negative or for uninfected births to HIV-positive
}
\details{
Assumes that children are born into the same subpopulation as their parents, that all children are born into the first risk state (unless otherwise specified) and the first non.hiv.subset (unless otherwise specified), and that distributions of sex by race are determined by the marginals on the given population
}
\seealso{
Other functions to create birth proportions arrays: 
\code{\link{create.birth.proportions.hiv.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.positive}()},
\code{\link{get.birth.proportions.to.hiv.negative.skeleton}()}

Other functions to specify births in the population: 
\code{\link{create.birth.proportions.hiv.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.positive}()},
\code{\link{get.birth.proportions.to.hiv.negative.skeleton}()},
\code{\link{set.birth.proportions.hiv.negative}()},
\code{\link{set.birth.proportions.hiv.positive}()},
\code{\link{set.birth.proportions.no.maternal.transmission}()},
\code{\link{set.fertility.hiv.negative}()},
\code{\link{set.fertility.hiv.positive}()},
\code{\link{set.fertility}()},
\code{\link{set.immigration.proportions}()}
}
\concept{functions to create birth proportions arrays}
\concept{functions to specify births in the population}
