% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{create.contact.array.from.marginals}
\alias{create.contact.array.from.marginals}
\title{Create a transmission contact array from one or more marginal contact arrays}
\usage{
create.contact.array.from.marginals(jheem, ...)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{...}{One or more arrays whose dimensions are a subset of: [age.from, race.from, subpopulation.from, sex.from, risk.from, age.to, race.to, subpopulation.to, sex.to, risk.to, non.hiv.subset.to]}
}
\description{
Create a transmission contact array from one or more marginal contact arrays
}
\seealso{
\code{\link{set.transmission.contact.array}}, \code{\link{get.contact.array.skeleton}}

Other functions for transmission contact arrays: 
\code{\link{get.contact.array.skeleton}()},
\code{\link{multiply.contact.array.from}()},
\code{\link{multiply.contact.array.to}()},
\code{\link{set.transmission.contact.array}()}
}
\concept{functions for transmission contact arrays}
\concept{functions to build arrays from one or more marginal arrays}
