% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{create.susceptibility.array.from.marginals}
\alias{create.susceptibility.array.from.marginals}
\title{Makes a susceptibility array by multiplying multiple marginal arrays}
\usage{
create.susceptibility.array.from.marginals(jheem, ...)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{...}{One or more arrays whose dimensions are a subset of the dimensions of the HIV-negative population: [age, race, subpopulation, sex, risk, non.hiv.subset]}
}
\description{
Susceptibility is a characteristic of the HIV-negative population, a multiplier indicating how likely an HIV-negative individual in each stratum is to aquire infection (per partnership) relative to a baseline. ie a value of 1 would not change the susceptibility from the global (baseline) rate
}
\seealso{
\code{\link{set.susceptibility}}, \code{\link{get.hiv.negative.population.skeleton}}
}
