% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{create.transmissibility.array.from.marginals}
\alias{create.transmissibility.array.from.marginals}
\title{Makes a transmissibility array by multiplying multiple marginal arrays}
\usage{
create.transmissibility.array.from.marginals(jheem, ...)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{...}{One or more arrays whose dimensions are a subset of the dimensions of the HIV-positive population: [age, race, subpopulation, sex, risk, continuum, cd4, hiv.subset]}
}
\description{
Transmissibilty is a characteristic of the HIV-positive population, a multiplier indicating how likely an HIV-positive individual in each stratum is to transmit infection (per partnership) relative to a baseline. ie a value of 1 would not change the transmissibility from the global (baseline) rate
}
\seealso{
\code{\link{set.transmissibility}}, \code{\link{get.hiv.positive.population.skeleton}}
}
