% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_population.R
\name{expand.population.to.general}
\alias{expand.population.to.general}
\title{Shoehorn a given population array into an array of the dimensions shared by both the HIV-negative and HIV-positive model populations}
\usage{
expand.population.to.general(jheem, population.to.expand)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{population.to.expand}{An array representing the population to expand. dimnames(population.to.expand) must be a subset of the dimnames of the HIV-negative population and of the dimnames of the HIV-positive population}
}
\description{
Returns an array with dimensions and dimnames shared by both the HIV-negative and HIV-positive populations, such that every element in the returned array is drawn from the population.to.expand array
}
\seealso{
\code{\link{expand.population.to.hiv.positive}}, \code{\link{expand.population.to.hiv.negative}}
}
