% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_population.R
\name{expand.population.to.hiv.positive}
\alias{expand.population.to.hiv.positive}
\title{Shoehorn a given population array into an array of the dimensions needed for the HIV-positive model population}
\usage{
expand.population.to.hiv.positive(jheem, population.to.expand)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{population.to.expand}{An array representing the population to expand. dimnames(population.to.expand) must be a subset of the dimnames of the HIV-positive population}
}
\description{
Returns an array with dimensions and dimnames matching the HIV positive population, such that every element in the returned array is drawn from the population.to.expand array
}
\seealso{
\code{\link{expand.population.to.hiv.negative}}
}
