% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_population.R
\name{expand.to.contact.array}
\alias{expand.to.contact.array}
\title{Shoehorn a given array into a transmission contact array}
\usage{
expand.to.contact.array(jheem, to.expand)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{to.expand}{An array representing the contact array to expand. dimnames(population.to.expand) must be a subset of the dimnames for a transmission contact array: [age.from, race.from, subpopulation.from, sex.from, risk.from, age.to, race.to, subpopulation.to, sex.to, risk.to, non.hiv.subset.to]}
}
\description{
Returns an array with dimensions and dimnames for a transmission contact array
}
