% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_population.R
\name{expand.to.new.infection.proportions}
\alias{expand.to.new.infection.proportions}
\title{Shoehorn a given population array into an array of the dimensions for new infection proportions}
\usage{
expand.to.new.infection.proportions(jheem, to.expand)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{to.expand}{An array representing the population to expand. dimnames(population.to.expand) must be a subset of the dimnames of the HIV-negative population and/or the dimnames of the HIV-positive population: [age, race, subpopulation, sex, risk, non.hiv.subset, continuum, cd4, hiv.subset]}
}
\description{
Returns an array with dimensions and dimnames for the new infection proportions array, such that every element in the returned array is drawn from the population.to.expand array
}
\details{
A new infection proportions array should have dimensions [age, race, subpopulation, sex, risk, non.hiv.subset, continuum, cd4, hiv.subset]. For each combination of age x race x subpopulation x sex x risk x non.hiv.subset, proportions[age,race,subpopulation,sex,risk,non.hiv.subset,,,] should sum to 1 and represent the proportions according to which new infections from the age x race x subpopulation x sex x risk x non.hiv.subset are distributed across strata of continuum x cd4 x hiv.subset
}
