% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{extract.diagnosed.hiv}
\alias{extract.diagnosed.hiv}
\title{Extract the subset of the HIV-positive population which has been diagnosed from JHEEM Results}
\usage{
extract.diagnosed.hiv(
  results,
  years = NULL,
  ages = NULL,
  races = NULL,
  subpopulations = NULL,
  sexes = NULL,
  risks = NULL,
  cd4s = NULL,
  hiv.subsets = NULL,
  keep.dimensions = "year",
  per.population = 1,
  use.cdc.categorizations = F
)
}
\arguments{
\item{results}{The results of a call to \code{\link{run.jheem}}}

\item{years, ages, races, subpopulations, risks, cd4s}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{keep.dimensions}{The names of which dimensions to marginalize over (all other dimensions will be summed out) - a subset of 'year', 'age', 'race', 'subpopulation', 'sex', 'risk', 'non.hiv.subset', 'continuum', 'cd4', and 'hiv.subset'. If passed null, defaults to keep the year dimension, and any other dimension for which more than one but less than all elements of that dimension are selected}

\item{per.population}{The unit of the denominator if the return value should be a proportion of all of those with HIV If NA, the absolute number is returned without dividing by the population size}
}
\description{
Extract the subset of the HIV-positive population which has been diagnosed from JHEEM Results
}
