% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{extract.incidence}
\alias{extract.incidence}
\title{Extract HIV incidence from JHEEM Results}
\usage{
extract.incidence(
  results,
  years = NULL,
  ages = NULL,
  races = NULL,
  subpopulations = NULL,
  sexes = NULL,
  risks = NULL,
  non.hiv.subsets = NULL,
  continuum = NULL,
  cd4s = NULL,
  hiv.subsets = NULL,
  keep.dimensions = NULL,
  include.hiv.positive.in.denominator = T,
  per.population = 1e+05,
  use.cdc.categorizations = F
)
}
\arguments{
\item{results}{The results of a call to \code{\link{run.jheem}}}

\item{years}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{ages}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{races}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{subpopulations}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{risks}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{non.hiv.subsets}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{continuum}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{cd4s}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{keep.dimensions}{The names of which dimensions to marginalize over (all other dimensions will be summed out) - a subset of 'year', 'age', 'race', 'subpopulation', 'sex', 'risk', 'non.hiv.subset', 'continuum', 'cd4', and 'hiv.subset'. If passed null, defaults to keep the year dimension, and any other dimension for which more than one but less than all elements of that dimension are selected}

\item{include.hiv.positive.in.denominator}{Whether to include HIV positive individuals in the denominator (only applies if per.population is set)}

\item{per.population}{The unit of the denominator if the return value should be a proportion of the population. If NA, the absolute number is returned without dividing by the population size}
}
\description{
Extract HIV incidence from JHEEM Results
}
