% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{extract.population.subset}
\alias{extract.population.subset}
\title{Extract a subset of the population from JHEEM Results}
\usage{
extract.population.subset(
  results,
  years = NULL,
  ages = NULL,
  races = NULL,
  subpopulations = NULL,
  sexes = NULL,
  risks = NULL,
  non.hiv.subsets = NULL,
  continuum = NULL,
  cd4s = NULL,
  hiv.subsets = NULL,
  include.hiv.positive = T,
  include.hiv.negative = T,
  keep.dimensions = NULL,
  denominator.dimensions = "year",
  per.population = NA,
  transformation.fn = NULL,
  use.cdc.categorizations = F
)
}
\arguments{
\item{results}{The results of a call to \code{\link{run.jheem}}}

\item{years, ages, races, subpopulations, risks, non.hiv.subsets, continuum, cd4s}{The elements of each of the possible result dimensions for which to extract results. If passed null, defaults to all elements of each dimension}

\item{include.hiv.positive, include.hiv.negative}{Whether to include the hiv.positive and hiv.negative subpopulations (at least one of these two must be set to true)}

\item{keep.dimensions}{The names of which dimensions to marginalize over (all other dimensions will be summed out) - a subset of 'year', 'age', 'race', 'subpopulation', 'sex', 'risk', 'non.hiv.subset', 'continuum', 'cd4', and 'hiv.subset'. If passed null, defaults to keep the year dimension, and any other dimension for which more than one but less than all elements of that dimension are selected}

\item{denominator.dimensions}{If the population should be normalized, which dimensions to use in the denominator (only applies if per.population is not NA). This must be a subset of the keep.dimensions. If NULL, the results are reported as fractions of the total population}

\item{per.population}{The unit of the denominator if the return value should be a proportion of the population. If NA, the absolute number is returned without dividing by the population size}
}
\description{
Extract a subset of the population from JHEEM Results
}
