% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{get.aging.skeleton.hiv.positive}
\alias{get.aging.skeleton.hiv.positive}
\title{Get a default array of aging rates for the HIV-positive population}
\usage{
get.aging.skeleton.hiv.positive(jheem)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}
}
\description{
Assumes that ages are distributed uniformly across the age bracket, such that the aging rate for an age bracket is 1 / the number of years in that age bracket
}
\seealso{
Other Functions to set aging rates: 
\code{\link{get.aging.skeleton.hiv.negative}()},
\code{\link{set.aging.hiv.negative}()},
\code{\link{set.aging.hiv.positive}()},
\code{\link{set.aging}()}
}
\concept{Functions to set aging rates}
