% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{get.birth.proportions.to.hiv.negative.skeleton}
\alias{get.birth.proportions.to.hiv.negative.skeleton}
\title{Create a skeleton Birth Proportions for births into the HIV-negative population (from either HIV-negative or HIV-positive parents)}
\usage{
get.birth.proportions.to.hiv.negative.skeleton(jheem, value = 0)
}
\arguments{
\item{jheem}{A JHEEM object}

\item{value}{The default value with which to populate the array}
}
\description{
Creates an empty birth proportions array for births into the HIV-negative population
}
\seealso{
Other functions to create birth proportions arrays: 
\code{\link{create.birth.proportions.from.population}()},
\code{\link{create.birth.proportions.hiv.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.positive}()}

Other functions to specify births in the population: 
\code{\link{create.birth.proportions.from.population}()},
\code{\link{create.birth.proportions.hiv.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.positive}()},
\code{\link{set.birth.proportions.hiv.negative}()},
\code{\link{set.birth.proportions.hiv.positive}()},
\code{\link{set.birth.proportions.no.maternal.transmission}()},
\code{\link{set.fertility.hiv.negative}()},
\code{\link{set.fertility.hiv.positive}()},
\code{\link{set.fertility}()},
\code{\link{set.immigration.proportions}()}
}
\concept{functions to create birth proportions arrays}
\concept{functions to specify births in the population}
