% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_skeletons.R
\name{get.contact.array.skeleton}
\alias{get.contact.array.skeleton}
\title{Get a skeleton contact array}
\usage{
get.contact.array.skeleton(
  jheem,
  value = 0,
  age = F,
  race = F,
  subpopulation = F,
  sex = F,
  risk = F,
  non.hiv.subset = F,
  age.from = age,
  race.from = race,
  subpopulation.from = subpopulation,
  sex.from = sex,
  risk.from = risk,
  age.to = age.from,
  race.to = race.from,
  subpopulation.to = subpopulation.from,
  sex.to = sex.from,
  risk.to = risk.from,
  non.hiv.subset.to = non.hiv.subset,
  all.dimensions = !age.from && !race.from && !subpopulation.from && !sex.from &&
    !risk.from && !age.to && !race.to && !subpopulation.to && !sex.to && !risk.to &&
    !non.hiv.subset.to
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{value}{The value to fill the array with. If value=='sequential', the array will be population with numbers 1:<size of the array>}

\item{age}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{race}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{subpopulation}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{sex}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{risk}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{non.hiv.subset}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{age.from}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{race.from}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{subpopulation.from}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{sex.from}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{risk.from}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{age.to}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{race.to}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{subpopulation.to}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{sex.to}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{risk.to}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{non.hiv.subset.to}{Indicators for whether these strata should be incorporated into the returned skeleton}

\item{all.dimensions}{If true, returns a contact array with all dimensions (age, race, subpopulation, sex, risk - both to and from) in it. If false, returns only the selected dimensions}
}
\description{
Gets an array, with all values set to value, with dimensions and names corresponding to the contact transmission array: [age from, race from, subpopulation from, sex from, risk from, age to, race to, subpopulation to, sex to, risk to, non-hiv subset to]
}
\seealso{
\code{\link{set.transmission.contact.array}}

Other functions for transmission contact arrays: 
\code{\link{create.contact.array.from.marginals}()},
\code{\link{multiply.contact.array.from}()},
\code{\link{multiply.contact.array.to}()},
\code{\link{set.transmission.contact.array}()}

Other functions to build population array skeletons: 
\code{\link{get.general.population.skeleton}()},
\code{\link{get.hiv.negative.population.skeleton}()},
\code{\link{get.hiv.positive.population.skeleton}()},
\code{\link{get.population.skeleton}()}
}
\concept{functions for transmission contact arrays}
\concept{functions to build population array skeletons}
