% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{get.hiv.negative.transition.array.skeleton}
\alias{get.hiv.negative.transition.array.skeleton}
\title{Create a skeleton transition array for the HIV-negative population}
\usage{
get.hiv.negative.transition.array.skeleton(
  jheem,
  transition.dimension,
  value = 0
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{transition.dimension}{The name of the dimension for which this is a transition array. Must be one of age, race, subpopulation, sex, risk, non.hiv.subset}

\item{value}{The value to fill the array with. If value=='sequential', the array will be population with numbers 1:<size of the array>}
}
\description{
Create a skeleton transition array for the HIV-negative population
}
