% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_skeletons.R
\name{get.hiv.positive.population.skeleton}
\alias{get.hiv.positive.population.skeleton}
\title{Get a skeleton HIV-positive population}
\usage{
get.hiv.positive.population.skeleton(jheem, value = 0)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{value}{The value to fill the array with. If value=='sequential', the array will be population with numbers 1:<size of the array>}
}
\description{
Gets an array, with all values set to value, with dimensions and names corresponding to the HIV-positive population for the specified JHEEM model
}
\seealso{
\code{\link{get.population.skeleton}}, \code{\link{get.general.population.skeleton}}, \code{\link{get.hiv.negative.population.skeleton}}

Other functions to build population array skeletons: 
\code{\link{get.contact.array.skeleton}()},
\code{\link{get.general.population.skeleton}()},
\code{\link{get.hiv.negative.population.skeleton}()},
\code{\link{get.population.skeleton}()}
}
\concept{functions to build population array skeletons}
