% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_skeletons.R
\name{get.population.skeleton}
\alias{get.population.skeleton}
\title{Get a skeleton for a population with indicated dimensions}
\usage{
get.population.skeleton(
  jheem,
  value = 0,
  age = F,
  race = F,
  subpopulation = F,
  sex = F,
  risk = F,
  non.hiv.subset = F,
  continuum = F,
  cd4 = F,
  hiv.subset = F,
  age.from = F,
  race.from = F,
  subpopulation.from = F,
  sex.from = F,
  risk.from = F,
  non.hiv.subset.from = F,
  continuum.from = F,
  cd4.from = F,
  hiv.subset.from = F,
  age.to = F,
  race.to = F,
  subpopulation.to = F,
  sex.to = F,
  risk.to = F,
  non.hiv.subset.to = F,
  continuum.to = F,
  cd4.to = F,
  hiv.subset.to = F
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{value}{The value to fill the array with. If value=='sequential', the array will be population with numbers 1:<size of the array>}

\item{age, race, subpopulation, sex, risk, non.hiv.subset, continuum, cd4, hiv.subset, age.from, race.from, subpopulation.from, sex.from, risk.from, non.hiv.subset.from, continuum.from, cd4.from, hiv.subset.from, age.to, race.to, subpopulation.to, sex.to, risk.to, non.hiv.subset.to, continuum.to, cd4.to, hiv.subset.to}{Indicators for whether these strata should be incorporated into the returned skeleton}
}
\description{
Gets an array, with all values set to value, with dimensions and names corresponding to the shared elements of both HIV-positive and HIV-negative populations (age, race, subpopulation, sex*risk) for the specified JHEEM model
}
\seealso{
\code{\link{get.population.skeleton}}, \code{\link{get.hiv.negative.population.skeleton}}, \code{\link{get.hiv.positive.population.skeleton}}

Other functions to build population array skeletons: 
\code{\link{get.contact.array.skeleton}()},
\code{\link{get.general.population.skeleton}()},
\code{\link{get.hiv.negative.population.skeleton}()},
\code{\link{get.hiv.positive.population.skeleton}()}
}
\concept{functions to build population array skeletons}
