% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get.seeded.initial.populations}
\alias{get.seeded.initial.populations}
\title{Get a pair of seeded initial states}
\usage{
get.seeded.initial.populations(
  jheem,
  nonhiv.init,
  cases.per = 1,
  seed.to.ages = jheem$age$labels[ceiling(length(jheem$age$labels)/2)],
  seed.to.races = jheem$race[1],
  seed.to.subpopulations = jheem$subpopulations[1],
  seed.to.sexes = jheem$sex[1],
  seed.to.risk.strata = jheem$risk.strata[1],
  seed.to.continuum.of.care = jheem$continuum.of.care[1],
  seed.to.cd4 = jheem$cd4.strata[1],
  seed.to.hiv.subset = jheem$hiv.subsets[1],
  seed.from.non.hiv.subset = jheem$nonhiv.subsets[1]
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{seed.to.ages}{The age strata into which cases will be seeded}

\item{seed.to.races}{The races into which cases will be seeded}

\item{seed.to.risk.strata}{The sex*risk strata into which cases will be seeded}

\item{cases.per.subpopulation}{Either (1) a numeric vector with length the number of subpopulations, where each value represents the number of cases to be seeded into each subpopulation, or (2) a single number, which will be the number of cases seeded into each subpopulation}
}
\description{
Seeds cases into an HIV-positive population, and subtracts those cases from the corresponding strata in the HIV-negative population
}
\seealso{
\code{\link{set.initial.state.hiv}}, \code{\link{get.hiv.negative.population.skeleton}}
}
