% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{get.susceptibility.array.skeleton}
\alias{get.susceptibility.array.skeleton}
\title{Get a skeleton susceptibility array}
\usage{
get.susceptibility.array.skeleton(jheem, value = 0)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{value}{The value to fill the array with}
}
\description{
Gets an array, with all values set to value, with dimensions and names corresponding to the HIV-negative population: [age, race, subpopulation, sex, risk, non.hiv.subset]
}
\seealso{
\code{\link{set.transmissibility}}

Other generators of population skeletons: 
\code{\link{get.transmissibility.array.skeleton}()}
}
\concept{functions for building susceptibility array}
\concept{generators of population skeletons}
