% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{get.transition.array.skeleton}
\alias{get.transition.array.skeleton}
\title{Create a skeleton transition array with arbitrary dimensions}
\usage{
get.transition.array.skeleton(
  jheem,
  transition.dimension,
  value = 0,
  age = F,
  race = F,
  subpopulation = F,
  sex = F,
  risk = F,
  non.hiv.subset = F,
  continuum = F,
  cd4 = F,
  hiv.subset = F
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{transition.dimension}{The name of the dimension for which this is a transition array.}

\item{value}{The value to fill the array with. If value=='sequential', the array will be population with numbers 1:<size of the array>}

\item{age, race, subpopulation, sex, risk, non.hiv.subset, continuum, cd4, hiv.subset}{Whether to include these dimensions in the skeleton array}
}
\description{
Create a skeleton transition array with arbitrary dimensions
}
