% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{get.transmissibility.array.skeleton}
\alias{get.transmissibility.array.skeleton}
\title{Get a skeleton transmissibility array}
\usage{
get.transmissibility.array.skeleton(jheem, value = 0)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{value}{The value to fill the array with}
}
\description{
Gets an array, with all values set to value, with dimensions and names corresponding to the HIV-positive population: [age, race, subpopulation, sex, risk, continuum, cd4, hiv.subset]
}
\seealso{
\code{\link{set.transmissibility}}

Other generators of population skeletons: 
\code{\link{get.susceptibility.array.skeleton}()}
}
\concept{functions for building transmissibility array}
\concept{generators of population skeletons}
