% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make.age.strata}
\alias{make.age.strata}
\title{Make Age Strata based on cutoffs}
\usage{
make.age.strata(age.cutoffs)
}
\arguments{
\item{age.cutoffs}{A numeric vector of ages, where age.cutoffs[1] is the lower bound of the first age bracket, age.cutoffs[length(age.cutoffs)] is the upper bound of the last age bracket}
}
\value{
a list with the following components: $labels, $lowers, $uppers, and $spans
}
\description{
makes length(age.cutoffs)-1 age strata
}
