% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{multiply.contact.array.from}
\alias{multiply.contact.array.from}
\title{Multiply a transmission contact array by rates for the 'from' population}
\usage{
multiply.contact.array.from(jheem, contact.array, from.multiplier)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{contact.array}{A contact array, with dimensions: [age.from, race.from, subpopulation.from, sex.from, risk.from, age.to, race.to, subpopulation.to, sex.to, risk.to, non.hiv.subset.to]}

\item{from.multiplier}{An array with dimensions matching the general population: [age, race, subpopulation, sex, risk]}
}
\description{
Fundamentally, a contact array can be thought of as taking the outer product of a 'from array' with a 'to array'. This function equates to multiplying the from array portion prior to taking the outer product with the to array portion
}
\details{
A new infection proportions array should have dimensions [age, race, subpopulation, sex, risk, non.hiv.subset, continuum, cd4, hiv.subset]. For each combination of age x race x subpopulation x sex x risk x non.hiv.subset, proportions[age,race,subpopulation,sex,risk,non.hiv.subset,,,] should sum to 1 and represent the proportions according to which new infections from the age x race x subpopulation x sex x risk x non.hiv.subset are distributed across strata of continuum x cd4 x hiv.subset
}
\seealso{
Other functions for transmission contact arrays: 
\code{\link{create.contact.array.from.marginals}()},
\code{\link{get.contact.array.skeleton}()},
\code{\link{multiply.contact.array.to}()},
\code{\link{set.transmission.contact.array}()}
}
\concept{functions for transmission contact arrays}
