% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{multiply.contact.array.to}
\alias{multiply.contact.array.to}
\title{Multiply a transmission contact array by rates for the 'to' population}
\usage{
multiply.contact.array.to(jheem, contact.array, to.multiplier)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{contact.array}{A contact array, with dimensions: [age.from, race.from, subpopulation.from, sex.from, risk.from, age.to, race.to, subpopulation.to, sex.to, risk.to, non.hiv.subset.to]}

\item{from.multiplier}{An array with dimensions matching the HIV-negative population: [age, race, subpopulation, sex, risk, non.hiv.subset]}
}
\description{
Fundamentally, a contact array can be thought of as taking the outer product of a 'from array' with a 'to array'. This function equates to multiplying the to array portion prior to taking the outer product with the from array portion
}
\seealso{
Other functions for transmission contact arrays: 
\code{\link{create.contact.array.from.marginals}()},
\code{\link{get.contact.array.skeleton}()},
\code{\link{multiply.contact.array.from}()},
\code{\link{set.transmission.contact.array}()}
}
\concept{functions for transmission contact arrays}
