% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeq.R
\name{run.jheem}
\alias{run.jheem}
\title{Run a JHEEM Model}
\usage{
run.jheem(
  jheem,
  start.year,
  end.year,
  keep.years = start.year:end.year,
  prior.run.results = NULL,
  verbose = T,
  print.warnings = T,
  max.run.time.seconds = Inf,
  atol = 1e-06,
  rtol = 1e-06
)
}
\arguments{
\item{jheem}{A fully initialized and set up jheem object}

\item{start.year, end.year}{The years between which to run the model (inclusive)}

\item{keep.years}{Which years to save results from}

\item{prior.run.results}{An object of class jheem.results, if starting values for this run should be taken from a prior run. If NULL, starting values will be taken from the initial state specified by the JHEEM object}

\item{verbose}{Whether to print detailed status updates}
}
\value{
An object of class jheem.results
}
\description{
Run a JHEEM Model
}
