% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.aging.hiv.negative}
\alias{set.aging.hiv.negative}
\title{Sets aging rates for HIV-negative population}
\usage{
set.aging.hiv.negative(jheem, aging.rates, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{time}{The time at which these mortality rates apply, if time-varying}
}
\description{
Sets aging rates for HIV-negative population
}
\seealso{
Other Functions to set aging rates: 
\code{\link{get.aging.skeleton.hiv.negative}()},
\code{\link{get.aging.skeleton.hiv.positive}()},
\code{\link{set.aging.hiv.positive}()},
\code{\link{set.aging}()}
}
\concept{Functions to set aging rates}
