% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.aging}
\alias{set.aging}
\title{Sets aging rates for both the HIV-negative and HIV-positive population}
\usage{
set.aging(jheem, aging.rates, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{time}{The time at which these mortality rates apply, if time-varying}

\item{aging}{An array of aging rates, whose dimensions are a subset of the dimensions shared by both the HIV-negative and HIV-positive populations (age, race, subpopulation, sex, risk factor). ***Rates should be per person per year}
}
\description{
Equivalent to calling set.aging.hiv.negative(jheem, aging.rates, time) and set.aging.hiv.positive(jheem, aging.rates, time). In general, aging rates can be approximated by the proportion of individuals in an age bracket who are in the last year of the age bracket
}
\seealso{
Other Functions to set aging rates: 
\code{\link{get.aging.skeleton.hiv.negative}()},
\code{\link{get.aging.skeleton.hiv.positive}()},
\code{\link{set.aging.hiv.negative}()},
\code{\link{set.aging.hiv.positive}()}
}
\concept{Functions to set aging rates}
