% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.birth.proportions.hiv.negative}
\alias{set.birth.proportions.hiv.negative}
\title{Sets the proportions according to which new births from HIV-negative parents are distributed across the HIV-negative population}
\usage{
set.birth.proportions.hiv.negative(jheem, proportions, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{proportions}{An array of birth proportions, typically created by calling \code{\link{create.birth.proportions.hiv.negative}}}

\item{time}{The time at which these proportions apply, if time-varying}
}
\description{
Sets the proportions according to which new births from HIV-negative parents are distributed across the HIV-negative population
}
\seealso{
Other functions to specify birth proportions: 
\code{\link{set.birth.proportions.hiv.positive}()},
\code{\link{set.birth.proportions.no.maternal.transmission}()},
\code{\link{set.immigration.proportions}()}

Other functions to specify births in the population: 
\code{\link{create.birth.proportions.from.population}()},
\code{\link{create.birth.proportions.hiv.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.positive}()},
\code{\link{get.birth.proportions.to.hiv.negative.skeleton}()},
\code{\link{set.birth.proportions.hiv.positive}()},
\code{\link{set.birth.proportions.no.maternal.transmission}()},
\code{\link{set.fertility.hiv.negative}()},
\code{\link{set.fertility.hiv.positive}()},
\code{\link{set.fertility}()},
\code{\link{set.immigration.proportions}()}
}
\concept{functions to specify birth proportions}
\concept{functions to specify births in the population}
