% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.birth.proportions.hiv.positive}
\alias{set.birth.proportions.hiv.positive}
\title{Sets the proportions according to which new births from HIV-positive are distributed across the HIV-negative and HIV-positive populations}
\usage{
set.birth.proportions.hiv.positive(
  jheem,
  fraction.births.infected,
  proportions.for.uninfected.births,
  proportions.for.infected.births = NULL,
  time = -Inf
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{fraction.births.infected}{An array that matches the HIV-positive population, indicating what fraction of offspring are born with HIV. Dimensions should be a subset of [age, race, subpopulation, sex, risk, continuum, cd4, hiv.subset]}

\item{proportions.for.uninfected.births, proportions.for.infected.births}{Arrays of birth proportions for uninfected and infected offpring, typically created by calling \code{\link{create.birth.proportions.hiv.positive.to.negative}} or \code{\link{create.birth.proportions.hiv.positive.to.positive}}}

\item{time}{The time at which these proportions apply, if time-varying}
}
\description{
Sets the proportions according to which new births from HIV-positive are distributed across the HIV-negative and HIV-positive populations
}
\seealso{
Other functions to specify birth proportions: 
\code{\link{set.birth.proportions.hiv.negative}()},
\code{\link{set.birth.proportions.no.maternal.transmission}()},
\code{\link{set.immigration.proportions}()}

Other functions to specify births in the population: 
\code{\link{create.birth.proportions.from.population}()},
\code{\link{create.birth.proportions.hiv.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.positive}()},
\code{\link{get.birth.proportions.to.hiv.negative.skeleton}()},
\code{\link{set.birth.proportions.hiv.negative}()},
\code{\link{set.birth.proportions.no.maternal.transmission}()},
\code{\link{set.fertility.hiv.negative}()},
\code{\link{set.fertility.hiv.positive}()},
\code{\link{set.fertility}()},
\code{\link{set.immigration.proportions}()}
}
\concept{functions to specify birth proportions}
\concept{functions to specify births in the population}
