% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.emigration.hiv.negative}
\alias{set.emigration.hiv.negative}
\title{Set emigration rates for HIV-negative population}
\usage{
set.emigration.hiv.negative(jheem, emigration, time = -Inf)
}
\arguments{
\item{mortality}{An array of mortality rates, whose dimensions are a subset of the HIV-negative population. ***Rates should be per person per year}
}
\description{
Set emigration rates for HIV-negative population
}
\seealso{
Other functions to specify emigration: 
\code{\link{set.emigration.hiv.positive}()},
\code{\link{set.emigration}()}
}
\concept{functions to specify emigration}
