% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.emigration.hiv.positive}
\alias{set.emigration.hiv.positive}
\title{Set emigration rates for HIV-positive population}
\usage{
set.emigration.hiv.positive(jheem, emigration, time = -Inf)
}
\arguments{
\item{emigration}{An array of emigration rates, whose dimensions are a subset of the dimensions of the HIV-negative population. ***Rates should be per person per year}
}
\description{
Set emigration rates for HIV-positive population
}
\seealso{
Other functions to specify emigration: 
\code{\link{set.emigration.hiv.negative}()},
\code{\link{set.emigration}()}
}
\concept{functions to specify emigration}
