% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.emigration}
\alias{set.emigration}
\title{Sets emigration rates for both the HIV-negative and HIV-positive population}
\usage{
set.emigration(jheem, emigration, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{emigration}{An array of emigration rates, whose dimensions are a subset of the dimensions shared by both the HIV-negative and HIV-positive populations (age, race, subpopulation, sex, risk factor). ***Rates should be per person per year}

\item{time}{The time at which these emigration rates apply, if time-varying}
}
\description{
Equivalent to calling set.emigration.hiv.negative(jheem, mortality, time) and set.emigration.hiv.positive(jheem, mortality, time)
}
\seealso{
Other functions to specify emigration: 
\code{\link{set.emigration.hiv.negative}()},
\code{\link{set.emigration.hiv.positive}()}
}
\concept{functions to specify emigration}
