% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.fertility.hiv.positive}
\alias{set.fertility.hiv.positive}
\title{Sets fertility rates for the HIV-positive population}
\usage{
set.fertility.hiv.positive(jheem, fertility, time = -Inf)
}
\arguments{
\item{jheem}{A JHEEM object}

\item{fertility}{An array of fertility rates, whose dimensions are a subset of [age, race, subpopulation, sex, risk, continuum, cd4, hiv.subset]. ***Rates should be per person per year}

\item{time}{The time at which these fertility rates apply, if time-varying}
}
\description{
*calling with a single scalar value for fertility would be equivalent to setting a birth rate (rather than a fertility rate) for the entire population
}
\seealso{
Other functions to specify births in the population: 
\code{\link{create.birth.proportions.from.population}()},
\code{\link{create.birth.proportions.hiv.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.negative}()},
\code{\link{create.birth.proportions.hiv.positive.to.positive}()},
\code{\link{get.birth.proportions.to.hiv.negative.skeleton}()},
\code{\link{set.birth.proportions.hiv.negative}()},
\code{\link{set.birth.proportions.hiv.positive}()},
\code{\link{set.birth.proportions.no.maternal.transmission}()},
\code{\link{set.fertility.hiv.negative}()},
\code{\link{set.fertility}()},
\code{\link{set.immigration.proportions}()}

Other functions to set fertility rates: 
\code{\link{set.fertility.hiv.negative}()},
\code{\link{set.fertility}()}
}
\concept{functions to set fertility rates}
\concept{functions to specify births in the population}
