% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.fixed.population.size}
\alias{set.fixed.population.size}
\title{Set the population to which to fix sizes}
\usage{
set.fixed.population.size(
  jheem,
  hiv.negative.population = jheem$initial.state.hiv.negative,
  hiv.positive.population = jheem$initial.state.hiv.positive,
  fixed.birth.proportions = create.birth.proportions.from.population(jheem,
    hiv.negative.population),
  time = -Inf
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{hiv.negative.population, hiv.positive.population}{Arrays representing the HIV-negative and HIV-populations to use in calculating the sizes for fixed strata}

\item{time}{The time at which the fixed population is pegged}
}
\description{
Must first call \code{\link{set.fixed.population.strata}}, which indicates which dimensions to fix population along. Calling set.fixed.population.size indicates what levels to fix the population to
}
\seealso{
\code{\link{set.fixed.population.strata}}, \code{\link{set.use.fixed.population}}

Other functions to fix population strata sizes: 
\code{\link{set.fixed.population.strata}()},
\code{\link{set.fixed.size.strata}()},
\code{\link{set.keep.strata.sizes.constant}()},
\code{\link{set.use.fixed.population}()}
}
\concept{functions to fix population strata sizes}
