% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.fixed.population.strata}
\alias{set.fixed.population.strata}
\title{Set which strata to use for fixed size populations}
\usage{
set.fixed.population.strata(
  jheem,
  fix.age = F,
  fix.race = F,
  fix.subpopulation = F,
  fix.sex = F,
  fix.risk = F
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{fix.age, fix.race, fix.subpopulation, fix.sex, fix.risk}{Boolean values indicating whether or not to fix these dimensions}
}
\description{
Set which strata to use for fixed size populations
}
\seealso{
\code{\link{set.use.fixed.population}}, \code{\link{set.fixed.population.size}}

Other functions to fix population strata sizes: 
\code{\link{set.fixed.population.size}()},
\code{\link{set.fixed.size.strata}()},
\code{\link{set.keep.strata.sizes.constant}()},
\code{\link{set.use.fixed.population}()}
}
\concept{functions to fix population strata sizes}
