% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.fixed.size.strata}
\alias{set.fixed.size.strata}
\title{Sets which dimensions define strata that will have sizes kept constant}
\usage{
set.fixed.size.strata(
  jheem,
  fix.age = F,
  fix.race = F,
  fix.subpopulation = F,
  fix.sex = F,
  fix.risk = F
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{fix.age, fix.race, fix.subpopulation, fix.sex, fix.risk}{Boolean values indicating whether or not these dimensions define the strata whose sizes are fixed}
}
\description{
Sets which dimensions define strata that will have sizes kept constant
}
\details{
Fixing strata sizes allows the JHEEM to adjust rates of change such that the overall size of the population within strata stays constant. The strata may be defined by any subset of the five dimensions shared by both HIV-negative and HIV-positive subpopulations: <age, race, subpopulation, sex, risk>.
Note that two separate functions must be called to use this ability: \code{\link{set.fixed.size.strata}} to indicte which dimensions define fixed strata and \code{\link{set.keep.strata.sizes.constant}} to actually indicate at what time to fix strata this way
}
\seealso{
Other functions to fix population strata sizes: 
\code{\link{set.fixed.population.size}()},
\code{\link{set.fixed.population.strata}()},
\code{\link{set.keep.strata.sizes.constant}()},
\code{\link{set.use.fixed.population}()}
}
\concept{functions to fix population strata sizes}
