% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.general.mortality.hiv.negative}
\alias{set.general.mortality.hiv.negative}
\title{Set mortality rates for HIV-negative population}
\usage{
set.general.mortality.hiv.negative(jheem, mortality, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{mortality}{An array of mortality rates, whose dimensions are a subset of the HIV-negative population. ***Rates should be per person per year}

\item{time}{The time at which these mortality rates apply, if time-varying}
}
\description{
Set mortality rates for HIV-negative population
}
\seealso{
\code{\link{set.hiv.specific.mortality}}, \code{\link{set.general.mortality}}, \code{\link{set.general.mortality.hiv.positive}}, \code{\link{expand.population.to.hiv.negative}}
}
