% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.general.mortality.hiv.positive}
\alias{set.general.mortality.hiv.positive}
\title{Set general mortality rates for HIV-positive population (not including HIV-specific mortality)}
\usage{
set.general.mortality.hiv.positive(jheem, mortality, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{mortality}{An array of mortality rates, whose dimensions are a subset of the dimensions of the HIV-negative population. ***Rates should be per person per year}

\item{time}{The time at which these mortality rates apply, if time-varying}
}
\description{
Set general mortality rates for HIV-positive population (not including HIV-specific mortality)
}
\seealso{
\code{\link{set.general.mortality}}, \code{\link{set.general.mortality.hiv.negative}}, \code{\link{expand.population.to.hiv.positive}}
}
