% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.general.mortality}
\alias{set.general.mortality}
\title{Sets mortality rates for both the HIV-negative and HIV-positive population}
\usage{
set.general.mortality(jheem, mortality, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{mortality}{An array of mortality rates, whose dimensions are a subset of the dimensions shared by both the HIV-negative and HIV-positive populations (age, race, subpopulation, sex, risk factor). ***Rates should be per person per year}

\item{time}{The time at which these mortality rates apply, if time-varying}
}
\description{
Equivalent to calling set.general.mortality.hiv.negative(jheem, mortality, time) and set.general.mortality.hiv.positive(jheem, mortality, time)
}
\seealso{
\code{\link{set.general.mortality.hiv.negative}}, \code{\link{set.general.mortality.hiv.positive}}
}
