% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{set.global.transmission.rate}
\alias{set.global.transmission.rate}
\title{Sets the global HIV transmission rate}
\usage{
set.global.transmission.rate(
  jheem,
  transmission.rate,
  transmission.route.names = jheem$transmission.routes,
  time = -Inf
)
}
\arguments{
\item{jheem}{An initialized JHEEM object}

\item{transmission.rate}{a single numeric value representing the global transmission rate, which can be used to tune the epidemic across all strata}

\item{transmission.route.names}{The names of the transmission routes to which the transmissibility should apply (a subset of the 'transmission.route.names' given to \code{\link{initialize.jheem}})}

\item{time}{The time at which this transmission rate applies, if time-varying}
}
\description{
Sets the global HIV transmission rate
}
\details{
The transmission rate (for one route of transmission) from stratum a (hiv-positive) to stratum b (hiv-negative) is a product of five components (four of which are set as parameters, and one of which - HIV prevalence - is determined by running the model):
\itemize{
 \item {\strong{A global transmission rate} - applies to all strata equally (\code{\link{set.global.transmission.rate}})}
 \item {\strong{Transmissibility} - all factors which depend only on the characteristics of the HIV-positive stratum a (\code{\link{set.transmissibility}})}
 \item {\strong{Contact rate} - all factors which depend on both the general characteristics of the HIV-positive stratum a (ie age, race, subpopulation, sex, risk factor) and the characteristics of the HIV-negative stratum b (\code{\link{set.transmission.contact.array}})}
 \item {\strong{Susceptibility} - all factors which depend only on the characteristics of the HIV-negative stratum b (\code{\link{set.susceptibility}})}
 \item {\strong{Prevalence of HIV} within the general characteristics of the HIV-positive stratum a (ie age, race, subpopulation, sex, risk factor). This is not set, but is estiated at model run time}
}
}
\seealso{
Other functions to specify transmission: 
\code{\link{set.new.infection.proportions}()},
\code{\link{set.susceptibility}()},
\code{\link{set.transmissibility}()},
\code{\link{set.transmission.contact.array}()}
}
\concept{functions to specify transmission}
