% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.hiv.specific.mortality}
\alias{set.hiv.specific.mortality}
\title{Set HIV-specific mortality rates for HIV-positive population}
\usage{
set.hiv.specific.mortality(jheem, mortality, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{mortality}{An array of mortality rates, whose dimensions are a subset of the dimensions of the HIV-negative population. ***Rates should be per person per year}

\item{time}{The time at which these mortality rates apply, if time-varying}
}
\description{
Set HIV-specific mortality rates for HIV-positive population
}
\seealso{
\code{\link{set.general.mortality}}, \code{\link{set.general.mortality.hiv.positive}}, \code{\link{expand.population.to.hiv.positive}}
}
