% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.immigration.rates}
\alias{set.immigration.rates}
\title{Sets immigration rates}
\usage{
set.immigration.rates(jheem, immigration.rates, is.relative = T, time = -Inf)
}
\arguments{
\item{jheem}{A JHEEM object}

\item{immigration.rates}{An array of immigration rates, whose dimensions are a subset of [age, race, subpopulation, sex, risk].}

\item{is.relative}{If true, the rates are interpreted as per person per year. If false, they are interpreted as absolute numbers}

\item{time}{The time at which these immigration rates apply, if time-varying}
}
\description{
Sets immigration rates
}
\details{
Immigration rates are interpreted similarly to birth rates: they are used to calculate the total number of immigrants, which are then distributed through either the HIV-negative or HIV-positive populations according to separately set proportions (set using \code{\link{set.immigration.proportions}})
}
\concept{functions to specify immigration}
