% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set.initial.population.hiv.negative}
\alias{set.initial.population.hiv.negative}
\title{Set the initial HIV-negative population}
\usage{
set.initial.population.hiv.negative(jheem, init)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{init}{An array representing the initial hiv-negative population}
}
\description{
Set the initial HIV-negative population
}
\seealso{
\code{\link{set.initial.state.hiv.positive}}, \code{\link{get.hiv.negative.population.skeleton}}
}
