% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set.initial.population.hiv.positive}
\alias{set.initial.population.hiv.positive}
\title{Set the initial HIV-positive population}
\usage{
set.initial.population.hiv.positive(jheem, init)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{init}{An array representing the initial hiv-positive population}
}
\description{
Set the initial HIV-positive population
}
\seealso{
\code{\link{set.initial.state.hiv.negative}}, \code{\link{get.hiv.positive.population.skeleton}}
}
