% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set.initial.populations}
\alias{set.initial.populations}
\title{Set the initial HIV-positive population}
\usage{
set.initial.populations(jheem, init.hiv.negative, init.hiv.positive)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{init.hiv.negative, init.hiv.positive}{Arrays representing the initial hiv-positive and hiv-negative populations}
}
\description{
A convenience function that bundles \code{\link{set.initial.population.hiv.positive}} and \code{\link{set.initial.population.hiv.negastive}} into one function call
}
\concept{functions to set model births}
