% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.keep.strata.sizes.constant}
\alias{set.keep.strata.sizes.constant}
\title{Sets whether to keep population strata sizes constant for a given time}
\usage{
set.keep.strata.sizes.constant(jheem, fix.strata.sizes, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{fix.strata.sizes}{A boolean indicated whether to use fixed strata sizes}

\item{time}{The time following which the setting applies}
}
\description{
Sets whether to keep population strata sizes constant for a given time
}
\details{
Fixing strata sizes allows the JHEEM to adjust rates of change such that the overall size of the population within strata stays constant. The strata may be defined by any subset of the five dimensions shared by both HIV-negative and HIV-positive subpopulations: <age, race, subpopulation, sex, risk>.
Note that two separate functions must be called to use this ability: \code{\link{set.fixed.size.strata}} to indicte which dimensions define fixed strata and \code{\link{set.keep.strata.sizes.constant}} to actually indicate at what time to fix strata this way
}
\seealso{
Other functions to fix population strata sizes: 
\code{\link{set.fixed.population.size}()},
\code{\link{set.fixed.population.strata}()},
\code{\link{set.fixed.size.strata}()},
\code{\link{set.use.fixed.population}()}
}
\concept{functions to fix population strata sizes}
