% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{set.track.hiv.specific.mortality}
\alias{set.track.hiv.specific.mortality}
\title{Set whether the JHEEM should track HIV-specific mortality}
\usage{
set.track.hiv.specific.mortality(
  jheem,
  track,
  dimensions.hiv.specific.mortality = c("age", "race", "subpopulation", "sex", "risk",
    "continuum", "cd4", "hiv.subset")
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{track}{Boolean indicating whether or not to track HIV-specific mortality}
}
\description{
Set whether the JHEEM should track HIV-specific mortality
}
\seealso{
Other function to set up JHEEM mortality: 
\code{\link{set.track.mortality}()},
\code{\link{set.track.overall.mortality}()}
}
\concept{function to set up JHEEM mortality}
