% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{set.track.incidence.dimensions}
\alias{set.track.incidence.dimensions}
\title{Set which dimensions to keep when tracking incidence}
\usage{
set.track.incidence.dimensions(
  jheem,
  dimensions = c("age", "race", "subpopulation", "sex", "risk", "non.hiv.subset",
    "continuum", "cd4", "hiv.subset")
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{dimensions}{The names of dimensions to keep when tracking incidence. Must be a subset of <'age','race','subpopulation','sex','risk','non.hiv.subset','continuum','cd4','hiv.subset'>}
}
\description{
Set which dimensions to keep when tracking incidence
}
