% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R, R/tracking.R
\name{set.track.mortality}
\alias{set.track.mortality}
\title{Set whether the JHEEM should track HIV-specific and overall mortality}
\usage{
set.track.mortality(
  jheem,
  track.hiv.specific.mortality = T,
  track.overall.hiv.positive.mortality = T,
  track.overall.hiv.negative.mortality = T,
  dimensions.hiv.specific.mortality = c("age", "race", "subpopulation", "sex", "risk",
    "continuum", "cd4", "hiv.subset"),
  dimensions.overall.hiv.positive.mortality = c("age", "race", "subpopulation", "sex",
    "risk", "continuum", "cd4", "hiv.subset"),
  dimensions.overall.hiv.negative.mortality = c("age", "race", "subpopulation", "sex",
    "risk", "non.hiv.subset")
)

set.track.mortality(
  jheem,
  track.hiv.specific.mortality = T,
  track.overall.hiv.positive.mortality = T,
  track.overall.hiv.negative.mortality = T,
  dimensions.hiv.specific.mortality = c("age", "race", "subpopulation", "sex", "risk",
    "continuum", "cd4", "hiv.subset"),
  dimensions.overall.hiv.positive.mortality = c("age", "race", "subpopulation", "sex",
    "risk", "continuum", "cd4", "hiv.subset"),
  dimensions.overall.hiv.negative.mortality = c("age", "race", "subpopulation", "sex",
    "risk", "non.hiv.subset")
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{track.hiv.specific.mortality, track.overall.hiv.positive.mortality, track.overall.hiv.negative.mortality}{Booleans indicating whether or not to track each type for mortality}

\item{dimensions.hiv.specific.mortality, dimensions.overall.hiv.positive.mortality}{The names of dimensions to keep when tracking hiv-positive mortality. Must be a subset of <'age','race','subpopulation','sex','risk','continuum','cd4','hiv.subset'>}

\item{dimensions.overall.hiv.negative.mortality}{The names of dimensions to keep when tracking HIV-negative mortality. Must be a subset of <'age','race','subpopulation','sex','risk','non.hiv.subset'>}
}
\description{
A convenience function that bundles \code{\link{set.track.hiv.specific.mortality}} and \code{\link{set.track.overall.mortality}} into one function call

A convenience function that bundles \code{\link{set.track.hiv.specific.mortality}} and \code{\link{set.track.overall.mortality}} into one function call
}
\seealso{
Other function to set up JHEEM mortality: 
\code{\link{set.track.hiv.specific.mortality}()},
\code{\link{set.track.overall.mortality}()}

Other function to set up JHEEM mortality: 
\code{\link{set.track.hiv.specific.mortality}()},
\code{\link{set.track.overall.mortality}()}
}
\concept{function to set up JHEEM mortality}
