% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{set.track.overall.mortality}
\alias{set.track.overall.mortality}
\title{Set whether the JHEEM should track general mortality}
\usage{
set.track.overall.mortality(
  jheem,
  track.hiv.positive,
  track.hiv.negative,
  dimensions.overall.hiv.positive.mortality = c("age", "race", "subpopulation", "sex",
    "risk", "continuum", "cd4", "hiv.subset"),
  dimensions.overall.hiv.negative.mortality = c("age", "race", "subpopulation", "sex",
    "risk", "non.hiv.subset")
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{track.hiv.positive, track.hiv.negative}{Boolean indicators for whether or not whether or not to track overall mortality for HIV-positive and HIV-negative}

\item{dimensions.overall.hiv.positive.mortality}{The names of dimensions to keep when tracking hiv-positive mortality. Must be a subset of <'age','race','subpopulation','sex','risk','continuum','cd4','hiv.subset'>}

\item{dimensions.overall.hiv.negative.mortality}{The names of dimensions to keep when tracking HIV-negative mortality. Must be a subset of <'age','race','subpopulation','sex','risk','non.hiv.subset'>}
}
\description{
Set whether the JHEEM should track general mortality
}
\details{
Overall mortality, for HIV-positive, is the sum of HIV-specific mortality and non-HIV mortality
}
\seealso{
Other function to set up JHEEM mortality: 
\code{\link{set.track.hiv.specific.mortality}()},
\code{\link{set.track.mortality}()}
}
\concept{function to set up JHEEM mortality}
