% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{set.tracked.transition.hiv.negative}
\alias{set.tracked.transition.hiv.negative}
\title{Specify an HIV-negative transition to track}
\usage{
set.tracked.transition.hiv.negative(
  jheem,
  name,
  transition.dimension,
  from,
  to,
  keep.dimensions = c("age", "race", "subpopulation", "sex", "risk", "non.hiv.subset"),
  age = NULL,
  race = NULL,
  subpopulation = NULL,
  sex = NULL,
  risk = NULL,
  non.hiv.subset = NULL
)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{name}{The name by which to refer to this tracked transition}

\item{transition.dimension}{The dimension to which this transition applies}

\item{from, to}{The names of the to and from states within the transition dimension that define the transition}

\item{keep.dimensions}{Which dimensions to keep in the tracked array (non-keep dimensions are marginalized out)}

\item{age, race, subpopulation, sex, risk, non.hiv.subset, cd4, continuum, hiv.subset}{The subset of the given dimensions for which to track transitions. If null, tracks the transitions across all values of the dimension}
}
\description{
Specify an HIV-negative transition to track
}
