% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{set.transition.array.hiv.positive}
\alias{set.transition.array.hiv.positive}
\title{Sets the array of transition rates between states (within strata of age and race) for the HIV-positive population}
\usage{
set.transition.array.hiv.positive(jheem, transition.array, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{transition.array}{A transition array}

\item{time}{The time at which these transition rates apply, if time-varying}
}
\description{
Sets the array of transition rates between states (within strata of age and race) for the HIV-positive population
}
\seealso{
Other JHEEM parameter setters: 
\code{\link{set.transition.array.hiv.negative}()}

Other functions for building transition arrays: 
\code{\link{set.transition.array.hiv.negative}()}
}
\concept{JHEEM parameter setters}
\concept{functions for building transition arrays}
