% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{set.transmission.contact.array}
\alias{set.transmission.contact.array}
\title{Sets the transmission contact array between HIV-negative and HIV-positive populations}
\usage{
set.transmission.contact.array(
  jheem,
  contact.array,
  transmission.route.names = jheem$transmission.routes,
  time = -Inf
)
}
\arguments{
\item{jheem}{An initialized JHEEM object}

\item{contact.array}{An contact array, with some subset of dimensions: [age.from, race.from, subpopulation.from, sex.from, risk.from, age.to, race.to, subpopulation.to, sex.to, risk.to, non.hiv.subset.to]. from->to follows the direction of HIV transmission}

\item{transmission.route.names}{The names of the transmission routes to which the transmissibility should apply (a subset of the 'transmission.route.names' given to \code{\link{initialize.jheem}})}

\item{time}{The time at which these susceptibilities apply, if time-varying}
}
\description{
A contact array is a function of both the hiv-infected population and hiv-negative population
}
\details{
The transmission rate (for one route of transmission) from stratum a (hiv-positive) to stratum b (hiv-negative) is a product of five components (four of which are set as parameters, and one of which - HIV prevalence - is determined by running the model):
\itemize{
 \item {\strong{A global transmission rate} - applies to all strata equally (\code{\link{set.global.transmission.rate}})}
 \item {\strong{Transmissibility} - all factors which depend only on the characteristics of the HIV-positive stratum a (\code{\link{set.transmissibility}})}
 \item {\strong{Contact rate} - all factors which depend on both the general characteristics of the HIV-positive stratum a (ie age, race, subpopulation, sex, risk factor) and the characteristics of the HIV-negative stratum b (\code{\link{set.transmission.contact.array}})}
 \item {\strong{Susceptibility} - all factors which depend only on the characteristics of the HIV-negative stratum b (\code{\link{set.susceptibility}})}
 \item {\strong{Prevalence of HIV} within the general characteristics of the HIV-positive stratum a (ie age, race, subpopulation, sex, risk factor). This is not set, but is estiated at model run time}
}
}
\seealso{
\code{\link{create.contact.array.from.marginals}}, \code{\link{get.contact.array.skeleton}}

Other functions for transmission contact arrays: 
\code{\link{create.contact.array.from.marginals}()},
\code{\link{get.contact.array.skeleton}()},
\code{\link{multiply.contact.array.from}()},
\code{\link{multiply.contact.array.to}()}

Other functions to specify transmission: 
\code{\link{set.global.transmission.rate}()},
\code{\link{set.new.infection.proportions}()},
\code{\link{set.susceptibility}()},
\code{\link{set.transmissibility}()}
}
\concept{functions for transmission contact arrays}
\concept{functions to specify transmission}
