% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth.R
\name{set.use.fixed.population}
\alias{set.use.fixed.population}
\title{Sets whether to use fixed population strata for a given time}
\usage{
set.use.fixed.population(jheem, fix.population, time = -Inf)
}
\arguments{
\item{jheem}{An object of \code{\link{jheem-class}}}

\item{fix.population}{A boolean indicated whether to use fixed population strata}

\item{time}{The time following which the setting applies}
}
\description{
Sets whether to use fixed population strata for a given time
}
\seealso{
\code{\link{set.fixed.population.strata}}, \code{\link{set.fixed.population.size}}

Other functions to fix population strata sizes: 
\code{\link{set.fixed.population.size}()},
\code{\link{set.fixed.population.strata}()},
\code{\link{set.fixed.size.strata}()},
\code{\link{set.keep.strata.sizes.constant}()}
}
\concept{functions to fix population strata sizes}
