% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{stratify.males.by.orientation}
\alias{stratify.males.by.orientation}
\title{Takes a population and splits males into heterosexual males and MSM}
\usage{
stratify.males.by.orientation(
  population.to.stratify,
  msm.multiplier = 1,
  heterosexual.multiplier = 1,
  male.name = "male",
  heterosexual.male.name = "heterosexual_male",
  msm.name = "msm"
)
}
\arguments{
\item{population.to.stratify}{A population where dimnames for one dimension includes male.name}

\item{msm.multiplier, heterosexual.multiplier}{The numbers by which to multiply the males in population.to.stratify to yield the corresponding populations in the stratified population}

\item{male.name}{The name given to males in a dimension of population.to.stratify}

\item{heterosexual.male.name, msm.name}{The names to be given to heterosexual males and MSM in the return value}
}
\description{
Returns an array with same dimensions and dimnames as population.to.stratify, except that the males have been split into heterosexual males and MSM
}
\seealso{
\code{\link{expand.population.to.hiv.negative}}
}
