% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{subset.jheem.results}
\alias{subset.jheem.results}
\title{Access a Subset of JHEEM Results by Year}
\usage{
\method{subset}{jheem.results}(results, years)
}
\arguments{
\item{results}{An object of class jheem.results to subset}

\item{years}{The years to subset}
}
\description{
Access a Subset of JHEEM Results by Year
}
