% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test.jheem}
\alias{test.jheem}
\title{Test JHEEM code to make sure that it is working correctly}
\usage{
test.jheem(error.behavior = c("stop", "browser", "continue")[1])
}
\arguments{
\item{error.behavior}{One of three choices: 'stop' (in which case the test function throws an error and stops), 'browser' (in which case the test function calls the browser), or 'continue' (in which case the function prints an error but continues running)}
}
\description{
Test JHEEM code to make sure that it is working correctly
}
\keyword{internal}
